🚚 Rename custom navlink horizontal navlink
