+++ frontend/app/routes/layouts/compose-stack-service-layout.tsx
@@ -20,7 +20,10 @@ import type { ComposeStackService } from "~/api/types";
 import { Code } from "~/components/code";
 import { CopyButton } from "~/components/copy-button";
 import { DeploymentStatusBadge } from "~/components/deployment-status-badge";
-import { type NavItem, NavLink } from "~/components/nav-link";
+import {
+  HorizontalNavLink,
+  type NavItem
+} from "~/components/horizontal-nav-link";
 import { StatusBadge } from "~/components/status-badge";
 import { SubmitButton } from "~/components/ui/button";
 import {
@@ -352,10 +355,10 @@ export default function ComposeStackServiceLayoutPage({
         >
           {navItems.map((item) => (
             <li key={item.href}>
-              <NavLink to={item.href} prefetch="viewport">
+              <HorizontalNavLink to={item.href} prefetch="viewport">
                 <span>{item.title}</span>
                 <item.icon size={15} className="flex-none" />
-              </NavLink>
+              </HorizontalNavLink>
             </li>
           ))}
         </ul>
